refactor(layout): move component into components folder and clean up global folder
