+++ src/components/Header/Header.styles.tsx
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
 
-import { media } from '../../global/MediaQueries';
+import { mediaQueries } from '../../theme/tokens';
 
 const HeaderWrapper = styled.div`
   background: ${props => props.theme.colors.background};
 
-  @media (${media.tablet}) {
-    background: ${props => props.theme.colors.background};
+  @media (${mediaQueries.tablet}) {
+    background: ${props => props.theme.colors.primary};
   }
 `;
