--- src/global/MediaQueries.ts
@@ -1,9 +1,0 @@
-import { breakpoints } from '../theme/tokens';
-
-const media = {
-  phone: `min-width: ${breakpoints[0]}`,
-  tablet: `min-width: ${breakpoints[1]}`,
-  desktop: `min-width: ${breakpoints[2]}`,
-};
-
-export { media };
