+++ src/pages/404.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Layout } from '../global/Layout';
+import { Layout } from '../components/Layout';
 
 const NotFoundPage = () => <Layout>404</Layout>;
