+++ src/theme/tokens/breakpoints.ts
@@ -1,1 +1,7 @@
 export const breakpoints: string[] = ['320px', '768px', '1024px'];
+
+export const mediaQueries = {
+  phone: `min-width: ${breakpoints[0]}`,
+  tablet: `min-width: ${breakpoints[1]}`,
+  desktop: `min-width: ${breakpoints[2]}`,
+};
