+++ src/theme/tokens/index.ts
@@ -1,6 +1,13 @@
 import { lightColors, darkColors } from './colors';
 import { space } from './space';
-import { breakpoints } from './breakpoints';
+import { breakpoints, mediaQueries } from './breakpoints';
 import { typographicScale } from './typographicScale';
 
-export { lightColors, darkColors, space, breakpoints, typographicScale };
+export {
+  breakpoints,
+  darkColors,
+  lightColors,
+  mediaQueries,
+  space,
+  typographicScale,
+};
